Move UnitStatsSerializationTest
The UnitStatsSerializationTest tests a class in the compiler.common project. It is unrelated to KGP.

Moving it helps because it moves the test closer to the actual class under test.
Currently, if the test is modified a KGP codeowner has to review it, which doesn't make sense and they may not have the right context.

To run the test this commit also adds build config for running tests in compiler.common project.

The test was originally added in `316af4b9959cb488d8d02b21e74c38d062706115` for KT-76498.
