+++ compiler/tests-integration/tests/org/jetbrains/kotlin/util/UnitStatsSerializationTest.kt
@@ -1,16 +1,15 @@
 /*
- * Copyright 2010-2025 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Copyright 2010-2026 JetBrains s.r.o. and Kotlin Programming Language contributors.
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.kotlin.gradle.unitTests.report
+package org.jetbrains.kotlin.util
 
 import com.google.gson.GsonBuilder
-import org.jetbrains.kotlin.util.*
-import kotlin.test.Test
 import java.time.LocalDateTime
 import java.time.ZoneOffset
 import kotlin.test.assertEquals
+import org.junit.jupiter.api.Test
 
 class UnitStatsSerializationTest {
     @Test
